requirements.txt
matplotlib==3.4.3
numpy==1.19.5
opencv-python-headless==4.5.3.56
pandas==1.3.3
Pillow==8.3.2
streamlit==1.0.0
tensorflow-cpu==2.6.0

